Refactor engine into smaller modules
