+++ py/scjson/activation.py
@@ -1,0 +1,72 @@
+"""
+Agent Name: activation-module
+
+Part of the scjson project.
+Developed by Softoboros Technology Inc.
+Licensed under the BSD 1-Clause License.
+
+Runtime activation records and transition specifications.
+"""
+
+from __future__ import annotations
+
+from enum import Enum
+from typing import Any, Dict, List, Optional
+
+from pydantic import BaseModel, ConfigDict, Field
+
+from .pydantic import Scxml, State, ScxmlParallelType, ScxmlFinalType
+
+SCXMLNode = State | ScxmlParallelType | ScxmlFinalType | Scxml
+
+
+class ActivationStatus(str, Enum):
+    """Lifecycle state for an activation."""
+
+    ACTIVE = "active"
+    FINAL = "final"
+
+
+class TransitionSpec(BaseModel):
+    """Simplified representation of a transition."""
+
+    event: Optional[str] = None
+    target: List[str] = Field(default_factory=list)
+    cond: Optional[str] = None
+
+
+class ActivationRecord(BaseModel):
+    """Runtime frame for an entered state/parallel/final element."""
+
+    model_config = ConfigDict(arbitrary_types_allowed=True)
+
+    id: str
+    node: SCXMLNode
+    parent: Optional["ActivationRecord"] = None
+    status: ActivationStatus = ActivationStatus.ACTIVE
+    local_data: Dict[str, Any] = Field(default_factory=dict)
+    children: List["ActivationRecord"] = Field(default_factory=list)
+    transitions: List["TransitionSpec"] = Field(default_factory=list)
+
+    def mark_final(self) -> None:
+        """Mark this activation as final and propagate upwards."""
+        self.status = ActivationStatus.FINAL
+        if self.parent and all(c.status is ActivationStatus.FINAL for c in self.parent.children):
+            self.parent.mark_final()
+
+    def add_child(self, child: "ActivationRecord") -> None:
+        """Append ``child`` to the list of children."""
+        self.children.append(child)
+
+    def is_active(self) -> bool:  # noqa: D401
+        """Return ``True`` while the activation is not finalised."""
+        return self.status is ActivationStatus.ACTIVE
+
+    def path(self) -> List["ActivationRecord"]:
+        """Return the path from the root activation to ``self``."""
+        cur: Optional["ActivationRecord"] = self
+        out: List["ActivationRecord"] = []
+        while cur:
+            out.append(cur)
+            cur = cur.parent
+        return list(reversed(out))
